Allow the repair to work in all cases

+++ setup.php
@@ -117,9 +117,14 @@ function plugin_flowview_check_upgrade() {
 				if (!cacti_sizeof($good)) {
 					cacti_log('NOTE: Updating unique key for ' . $t['TABLE_NAME'], false, 'FLOWVIEW');
 
-					db_execute('ALTER TABLE ' . $t['TABLE_NAME'] . '
-						DROP INDEX `keycol`,
-						ADD UNIQUE INDEX `keycol` (`listener_id`, `src_addr`, `src_port`, `src_if`,`dst_addr`,`dst_port`, `dst_if`, `start_time`, `end_time`)');
+					if (db_index_exists($t['TABLE_NAME'], 'keycol')) {
+						db_execute('ALTER TABLE ' . $t['TABLE_NAME'] . '
+							DROP INDEX `keycol`,
+							ADD UNIQUE INDEX `keycol` (`listener_id`, `src_addr`, `src_port`, `src_if`,`dst_addr`,`dst_port`, `dst_if`, `start_time`, `end_time`)');
+					} else {
+						db_execute('ALTER TABLE ' . $t['TABLE_NAME'] . '
+							ADD UNIQUE INDEX `keycol` (`listener_id`, `src_addr`, `src_port`, `src_if`,`dst_addr`,`dst_port`, `dst_if`, `start_time`, `end_time`)');
+					}
 				}
 			}
 		}
